// src/features/compliance/interfaces.tsx
export interface IResponse<Type> {
	results: Type[];
}
// ...
export interface ReportReasonEntityProperties {
	reasonText: string;
	reportReasonCategoryId: string;
	reportReasonCategory: string;
	reportReasonPriorityId: string;
	reportReasonPriority: string;
	mlsRuleId: string;
	mlsRule: string;
	createdBy: string;
	createdOn: string;
	id: string;
	modifiedBy: string;
	modifiedOn: string;
}

export interface IAudienceMember {
	contactId: string;
	loginId: string;
	firstName: string;
	lastName: string;
	emailAddress: string;
	aorShortName: string;
	applicationShortName: string;
	mlsShortName: string;
	memberStatus: string;
	memberType: string;
	aorName: string;
	applicationName: string;
	mlsName: string;
	memberStatusName: string;
	memberTypeName: string;
	searchIndex: string;
	contactJObject: any;
	createdOn: string;
	id: number;
	modifiedOn: string;
}